# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for wallet and transaction management with JWT auth and RBAC.",
        "title": "ExpenseTracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Shows user's transactions with optional filters like category and from and to dates, where the displayed amount is determined by page and limit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieves user's transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Which page of the transactions list to show",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "How many transactions to show in each page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Which category of transactions to show (case-insensitive)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Which date to start showing transactions after in RFC3339",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Which date to start showing transactions before in RFC3339",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "failed to parse a query param",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "401": {
                        "description": "invalid authorization",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401"
                        }
                    },
                    "500": {
                        "description": "failed to correctly query the database or parse claims",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
        "/transactions/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Shows user's transactions grouped by category and formatted to be concise and nice to read",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieves a summary of user's transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SummaryResult"
                        }
                    },
                    "400": {
                        "description": "failed to parse a query param",
                        "schema": {
                            "$ref": "#/definitions/dto.Error400"
                        }
                    },
                    "401": {
                        "description": "invalid authorization",
                        "schema": {
                            "$ref": "#/definitions/dto.Error401"
                        }
                    },
                    "500": {
                        "description": "failed to correctly query the database or parse claims",
                        "schema": {
                            "$ref": "#/definitions/dto.Error500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.Error400": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "dto.Error401": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "missing claims"
                }
            }
        },
...
        "dto.Error500": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "failed to query database"
                }
            }
        },
...
        "handler.CategorySummary": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handler.SummaryResult": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CategorySummary"
                    }
                },
                "period": {
                    "type": "string",
                    "example": "2026-08"
                },
                "total": {
                    "type": "integer",
                    "example": 20000
                }
            }
        },
        "model.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-08-13T22:08:50.997585+03:00"
                },
                "id": {
                    "type": "integer",
                    "example": 4
                },
                "note": {
                    "type": "string"
                },
                "related_wallet_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-08-13T22:08:50.997585+03:00"
                },
                "wallet_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}